+++ include/NeoFOAM/fields/boundaryFields.hpp
@@ -115,6 +115,15 @@ public:
 
     const Executor& exec() { return exec_; }
 
+    /**
+     * @brief Get the range for a given patchId
+     * @return The number of boundary faces.
+     */
+    std::pair<localIdx, localIdx> range(localIdx patchId) const
+    {
+        return {offset_.data()[patchId], offset_.data()[patchId + 1]};
+    }
+
 private:
 
     Executor exec_;                        ///< The executor on which the field is stored
@@ -130,4 +139,4 @@ private:
     int nBoundaryFaces_;                   ///< The number of boundary faces.
 };
 
-} // namespace NeoFOAM
+}